winning-project/RHXQC/sourcecode/4-Platform/recommendation-system-with-word2vec-and-streamlit-master/requirements.txt
streamlit
pytest-shutil
pandas
numpy
tqdm
sklearn
Google-Images-Search
matplotlib
google_images_download
pillow
python-resize-image
joblib
gensim
